[BAG] fix for delay in title/favicon loading after opening window from bag
When a saved window from the bag is opened, a new physical window is created
and filled with the saved tabs. Simultaneously, a WinFrame object is created
in TabSense UI and is populated with Tab objects. The regular Tab object gets
its title and favIconUrl from the physical tab object. However when they are
freshly being created, their title and favIcon URLs are not ready. This fix
instead uses the title and favIconUrl stored in bag to show in TabSense UI.