A bunch of changes to get saving and retriving windows from bag working
The window and its tabs are marked 'saved' in the database. when opening it
(unsaving), the window and tab information is read, and removed from DB.
This info is used to fork new window and tabs. The onCreated event handlers for
this window and tabs create new rows in the database, which have
all the same info except the tabId (tid).

Also the parent and depth information is ignored, because it is
based on old tab IDs, so it will be difficult to recreate the
same hierarchy.