[ATTIC] backend
when tabs are collapsed or expanded in front end, move the
hidden tabs to and from attic.

Attic is a special window which we use to temporarily hold
the tabs, so that the tab clutter in main window reduces.

Ideally the collapsed tabs should be made invisible, but chrome
extension API doesn't provide any call to do that.