Change what happen to parent tab when it's collapsed
In last couple of commits:
The parent tab was also moved into the attic when collapsed. This led to problem
when the parent tab was the last tab in window. Moving it to attic would close
the window. To avoid that, we used collapse.html and that brought lot of more
code. The parent tab was moved to attic to save us from a possible user action.
What if user creates another child while the parent tab is collapsed. If the
parent tab is not present, he won't be able to create a child at all.

In this commit:
The parent tab is not moved to attic. This lets us get rid of collapse.html
and all other code it requires. If the user creates children from a collapsed
tab, we immediately expand the tab. This seems more intuitive thing to do
and also saves us hackish collapse.html code.