Handle Attic window across restarts
This fix, recognizes the attic window, by looking for attic tab in new windows
This window is not compared with DB windows. Also hidden tabs are ommitted from
DB windows, so that they match real windows (hidden tabs are not in real windows
but are in attic window).

Go through new tab IDs of tabs in attic window and update the hidden tabs' IDs
with them.